put legacy stuff into legacy, expose `next` api
